refactor(FE): `useTokenBalance` hook
